Long Int

+++ app/src/main/java/com/tung/coffeeorder/Reward.kt
@@ -7,10 +7,10 @@ class Reward(order: Order) {
         return order
     }
 
-    fun calculateBonusPoint(): Int{
-        var res=0
+    fun calculateBonusPoint(): Long{
+        var res=0L
         for (coffeeInCart in order.getCart()){
-            res+=(coffeeInCart.getPrice()/1000) as Int
+            res+=(coffeeInCart.getPrice()/1000)
         }
         return res
     }
